backend/requirements.txt
alembic~=1.14.0
python-dotenv~=1.0.1
psycopg2-binary
loguru
python-jose~=3.3.0
passlib[bcrypt]~=1.7.4
fastapi[standard]~=0.115.6
sqlalchemy~=2.0.36
starlette~=0.41.3
cryptography~=44.0.0
pydantic~=2.10.4